Refactors decorator into its own file, fixes small logic bug in getAllReservations
